chatbot/src/chatbot/main3.py
# mix for both agent and chainlit and include history 

import os
from dotenv import load_dotenv
from typing import cast
import chainlit as cl
from agents import Agent, Runner, AsyncOpenAI, OpenAIChatCompletionsModel
from agents.run import RunConfig


load_dotenv()

gemini_api_key = os.getenv("GEMINI_API_KEY")

# Check if the API key is present; if not, raise an error
if not gemini_api_key:
    raise ValueError("GEMINI_API_KEY is not set. Please ensure it is defined in your .env file.")


#Reference: https://ai.google.dev/gemini-api/docs/openai
provider = AsyncOpenAI(
    api_key=gemini_api_key,
    base_url="https://generativelanguage.googleapis.com/v1beta/openai/",
)

model = OpenAIChatCompletionsModel(
    model="gemini-2.0-flash",
    openai_client=provider
)

config = RunConfig(
    model=model,
    model_provider=provider,
    tracing_disabled=True
)


agent: Agent = Agent(name="Assistant", instructions="You are a helpful assistant", model=model)


@cl.on_chat_start
async def handal_chat_start():
    cl.user_session.set("history", [])
    await cl.Message(content="Welcome to the Panaversity AI Assistant! How can I help you today?").send()


@cl.on_message
async def handle_message(message: cl.Message):
    history = cl.user_session.get("history")
    history.append({"role":"user", "content": message.content})
    result = await Runner.run(agent, input=history, run_config=config) # here place await and then Runner.run instead of Runner.sync
    history.append({"role":"user", "content": result.final_output})
    cl.user_session.set("history", history)
    await cl.Message(content=result.final_output).send()

